refactor(GCFLogger-test-helpers): rename to test-helpers.ts
